Add readme, make script with with longer videos

+++ src/scripts/organize_tournament_videos.py
@@ -3,43 +3,77 @@ import json
 import os
 import shutil
 
-from utils.files import get_video_length, list_files_sorted_by_date
+from utils.files import get_video_length, list_files_of_type_sorted_by_date
 from utils.google_sheet_reader import get_google_sheet_data, parse_schedule
 from utils.utils import log, format_team_name_for_filename
 
 # Rename the videos in the directory to the following format: "Home Team vs. Away Team.mp4"
 # directory_name: the name of the directory containing the videos
-def rename_videos(directory_name, games_list):
+def rename_videos(directory_name, games_list, dry_run=False):
     output_directory = f'{directory_name}/processed_videos'
     if not os.path.exists(output_directory):
         os.makedirs(output_directory)
 
-    videos = get_likely_ordered_game_filenames(directory_name)
     video_paths = []
-    for idx, video in enumerate(videos):
-        game = games_list[idx]
+    videos = get_likely_ordered_game_filenames(directory_name)
+    idx_offset = 0
+    for game_idx, video_tuple in enumerate(videos):
+        if game_idx == 5: # List any missed games here
+            idx_offset = 1
+        game = games_list[game_idx + idx_offset] # In case any games are missed
         home_team = game["home_team"]
         away_team = game["away_team"]
         round = game["round"]
-        new_video_name = f"{format_team_name_for_filename(home_team)}_{format_team_name_for_filename(away_team)}_round_robin_round_{round}.mp4"
-        new_video_path = os.path.join(output_directory, new_video_name)
-        old_video_path = os.path.join(directory_name, video)
-        shutil.copy(old_video_path, new_video_path)
-        log(f"Copied {video} to {new_video_path}")
-        game["video_path"] = new_video_path
-        video_paths.append(new_video_path)
+
+
+        for idx, video in enumerate(video_tuple, start=1):
+            new_video_name = f"{format_team_name_for_filename(home_team)}_{format_team_name_for_filename(away_team)}_round_robin_round_{round}_part_{idx}.mp4"
+            new_video_path = os.path.join(output_directory, new_video_name)
+            old_video_path = os.path.join(directory_name, video)
+
+            if not dry_run:
+                shutil.copy(old_video_path, new_video_path)
+            log(f"Copied {video_tuple} to {new_video_path}")
+            game["video_path"] = new_video_path
+            video_paths.append(new_video_path)
+
     return output_directory, video_paths
 
 
 # Pull all videos in the directory greater than five minutes long
 def get_likely_ordered_game_filenames(directory_name):
     game_video_filenames = []
-    for video in list_files_sorted_by_date(directory_name):
-        if video.endswith(".mp4"):
-            video_path = os.path.join(directory_name, video)
-            video_length = get_video_length(video_path)
-            if video_length >= 300:
-                game_video_filenames.append(video)
+    filenames_in_this_recording = []
+    for video in list_files_of_type_sorted_by_date(directory_name):
+            
+        # Files are named like GX, a 2 digit number, then a 4 digit number.
+        # When we go from one video to the next, if the 4 digit number stays the same
+        # and the 2 digit number increments, the two videos are one recording split into two files.
+        # When the 4-digit number increments, it's a new game.
+        video_prefix = video[2:4]
+        video_suffix = video[4:8]
+
+        if filenames_in_this_recording:
+            if video_suffix == filenames_in_this_recording[-1][4:8]: # Still creating the list for this recording
+                filenames_in_this_recording.append(video)
+                log(f"Appending {video} to filenames_in_this_recording")
+            else: # This is a new recording, so we need to update the game_video_filenames with a tuple of each filename in this recording
+                # turn the list of filenames into a tuple
+                game_video_filenames.append(tuple(filenames_in_this_recording))
+                filenames_in_this_recording = [video]
+                log(f"New recording: {video}")
+                log(f"Adding {filenames_in_this_recording} to game_video_filenames")
+                log(f"Game video filenames: {game_video_filenames}")
+        else:
+            filenames_in_this_recording.append(video)
+            log(f"First video: {video}")
+    # Add the last recording
+    game_video_filenames.append(tuple(filenames_in_this_recording))
+
+    # Check through the videos and if the tuple has just one item AND that video is less than five minutes long, remove it from the list
+    game_video_filenames = [video for video in game_video_filenames if len(video) > 1 or get_video_length(os.path.join(directory_name, video[0])) > 300]
+
+
     log(f"Game video filenames: {game_video_filenames}")
     return game_video_filenames
 
@@ -62,9 +96,9 @@ def create_metadata_file(schedule, output_path, video_paths):
         json.dump(metadata, f)
     log(f"Metadata file created at {output_path}/metadata.json")
 
-def run(directory_name, ordered_games):
+def run(directory_name, ordered_games, dry_run=False):
     # Rename the videos in the directory
-    output_path, video_paths = rename_videos(directory_name, ordered_games)
+    output_path, video_paths = rename_videos(directory_name, ordered_games, dry_run)
     create_metadata_file(ordered_games, output_path, video_paths)
     
 
@@ -72,6 +106,8 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Handle tournament videos.')
     parser.add_argument('directory_name', type=str, help='The name of the directory containing the videos')
     parser.add_argument('--court', type=int, help='The court number')
+    parser.add_argument('--dry-run', action='store_true', help='If true, do not copy the files, just print the output')
+
     args = parser.parse_args()
 
     sheet_data = get_google_sheet_data()
@@ -89,4 +125,4 @@ if __name__ == '__main__':
 
     opening_screens_output_path = f"{output_path}/opening_screens"
 
-    run(args.directory_name, ordered_games)
+    run(args.directory_name, ordered_games, args.dry_run)
